Move shared log plumbing into rise-backend-core
src/server/deployment/logs.rs interleaved four runtime log backends —
Kubernetes, Docker, Loki, None — with the plumbing all of them share, in one
4,241-line file. The Kubernetes backend cannot leave the root crate while
that plumbing lives above it and the other backends depend on the same code.

rise-backend-core::logs now owns the shared half: the query and event types,
the RuntimeLogBackend trait, the container-stream merge with its bounded
newest-N heap, cursor encoding and stable line identity, and the level
classifier. logs.rs re-exports it, so every `logs::LogQuery`-style path in
handlers and the CloudWatch backend keeps resolving.

Two names claimed Kubernetes ownership of generic code — cloudwatch.rs was
already importing both. `classify_k8s_line` becomes `classify_log_line` and
`KUBERNETES_LEVELS` becomes `HEURISTIC_LEVELS`, described by what they do
rather than by the first backend that needed them.

Eight tests covering the moved code move with it, so core is testable on its
own rather than only through the root crate. The rest stay: they cover
Kubernetes tail-offset math and Loki selectors, which are not going anywhere
yet. Whole-suite count is unchanged at 1,515.

logs/cloudwatch.rs is the ECS log backend and still sits in the root crate.
Now that rise-backend-core::logs exists, moving it into rise-backend-ecs is
a pure move; tracked as a follow-up rather than folded in here.
